--- test/mininet/gg/basic.js
@@ -1,4 +1,0 @@
-'use strict'
-process.env.NODES = 4
-process.env.RTS = 1000
-require('./horde')
