+++ test/mininet/package.json
@@ -7,7 +7,6 @@
     "grenache-nodejs-http": "^0.7.1",
     "grenache-nodejs-link": "bitfinexcom/grenache-nodejs-link",
     "lodash.mergewith": "^4.6.2",
-    "mininet": "git+https://github.com/davidmarkclements/mininet.git",
     "tapenet": "^1.4.0"
   },
   "scripts": {
